+++ ext/num4hypothtst/hypothtst/GrubbsTest.java
@@ -2,10 +2,10 @@ package hypothtst;
 
 import org.apache.commons.math3.distribution.TDistribution;
 // グラブス・スミルノフの外れ値の検定
-public class GrubbsTest {
-    private static GrubbsTest grubbs = new GrubbsTest();
-    public static GrubbsTest getInstance() {
-        return grubbs;
+public class GrubbsTest implements HypothTest2 {
+    private static HypothTest2 hypothTest2 = new GrubbsTest();
+    public static HypothTest2 getInstance() {
+        return hypothTest2;
     }
     public boolean twoSideTest(double statistic, int n, double a) {
         double r_val = calcGnValue(n, a / 2.0);
@@ -15,6 +15,7 @@ public class GrubbsTest {
         double r_val = calcGnValue(n, a);
         return evaluation(statistic, r_val);
     }
+
     private double calcGnValue(int n, double a) {
         TDistribution tDist = new TDistribution(n);
         double t = tDist.inverseCumulativeProbability(a / n);