+++ ext/num4hypothtst/hypothtst/HypothTest2.java
@@ -1,7 +1,8 @@
 package hypothtst;
 
 // 仮設検定
-interface HypothTest {
+interface HypothTest2 {
     boolean twoSideTest(double statistic, int n, double a);
+    boolean oneSideTest(double statistic, int n, double a);
 }