+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Unreleased
 
+## [0.0.6] - 2023-12-27
+
+### delete
+- del function of wilcoxon.
+
+### add
+- add function of decorrelation.
+- add function of populationCorre.
+
 ## [0.0.5] - 2023-12-19
 
 ### add